Fixes and updates, version 0.8.28 (#18)
* Add 0.6.10 through 0.6.12 to download.sh
* Update 0.8.24 hash to match universal binary
I'm not 100% sure if this is what's desired, but I noticed that the cached version must be the old non-universal binary for 0.8.24 instead of the universal binary. When running a full rebuild, this is the hash I get.
* Add wget and jq to download.sh shell
I'm adding these because they're not part of POSIX and might not be installed in every environment.
* Use realpath before running nix hash file
This seems to be necessary for my system. nix hash file gives an error: path '.../bin/macos-aarch64/solc-0.8.x' is a symlink, which this change fixes.
* Add 0.8.28

---------

Co-authored-by: Miao, ZhiCheng <zhicheng.miao@gmail.com>